+++ src/appleseed/renderer/modeling/camera/orthographiccamera.cpp
@@ -157,7 +157,7 @@ namespace
 
         virtual bool connect_vertex(
             SamplingContext&    sampling_context,
-            const double        time,
+            const float         time,
             const Vector3d&     point,
             Vector2d&           ndc,
             Vector3d&           outgoing,
@@ -205,7 +205,7 @@ namespace
         }
 
         virtual bool project_segment(
-            const double        time,
+            const float         time,
             const Vector3d&     a,
             const Vector3d&     b,
             Vector2d&           a_ndc,
